edit: refactor
