+++ go.mod
@@ -3,7 +3,7 @@ module go.aporeto.io/manipulate
 go 1.13
 
 require (
-	go.aporeto.io/elemental v1.100.1-0.20200731065958-3c0219b4b984
+	go.aporeto.io/elemental v1.100.1-0.20200918171102-0bf423f6b322
 	go.aporeto.io/wsc v1.36.1-0.20200731064733-e5c5718dc58f
 )
