+++ manipmongo/compiler_test.go
@@ -9,7 +9,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package compiler
+package manipmongo
 
 import (
 	"strings"
