+++ manipmongo/helpers.go
@@ -18,23 +18,10 @@ import (
 	"time"
 
 	"github.com/globalsign/mgo"
-	"github.com/globalsign/mgo/bson"
 	"go.aporeto.io/elemental"
 	"go.aporeto.io/manipulate"
 	"go.aporeto.io/manipulate/internal/backoff"
-	"go.aporeto.io/manipulate/manipmongo/internal/compiler"
-	compilerOpts "go.aporeto.io/manipulate/manipmongo/options/compiler"
 )
-
-type compilerConfig struct {
-	TranslateKeysFromSpec bool
-	AttrSpecs             map[string]elemental.AttributeSpecification
-}
-
-// CompileFilter compiles the given manipulate filter into a raw mongo filter.
-func CompileFilter(f *elemental.Filter, opts ...compilerOpts.CompilerOption) bson.D {
-	return compiler.CompileFilter(f, opts...)
-}
 
 // DoesDatabaseExist checks if the database used by the given manipulator exists.
 func DoesDatabaseExist(manipulator manipulate.Manipulator) (bool, error) {
